moved shared comps to /app
