+++ src/app/(app)/lists/[id]/items/ItemDndContext.tsx
@@ -17,8 +17,8 @@ import { useAuth } from "app/authContext";
 import { addDoc, deleteDoc, updateDoc } from "firebase/firestore";
 import { useState } from "react";
 import { useItems } from "app/(app)/lists/[id]/itemsContext";
-import { Loading } from "app/(app)/shared/Loading";
-import { Error } from "app/(app)/shared/Error";
+import { Loading } from "app/shared/Loading";
+import { Error } from "app/shared/Error";
 import { Item as ItemType, Section, AdminList } from "types/types";
 import { createItemData } from "db/factory";
 import { itemsOfList, itemsOfSection } from "db/useDb";
